Optimize the article selection code.
Heavily reduce network communication on QUICK mode. Add an option to
extract only the last edit time instead of the whole page (suboptimal
for now). Repack the structure of the page data before going through
files.

Results show a 3-fold improvement in QUICK mode and a 5-10% improvement
in NORMAL mode.