Implement an error-resilient way to find the code.
The new code uses regex to find the code. If several
codes are found, it tries to find the one marked with
the template. If this does not succeed, it tries to
find the one from the infobox.

If several codes are still found, it uses a simple
(database-specific) algorithm to deterine if they are
submonuments from the same ensemble.