Add the option to force the update of the fields
This replaces the older, hardcoded logic on whether
to update or not a non-empty field. It is meant to
be used in conjunction with the -update* params.

Also slightly modify the logic to favor images from
articles if a single commons image is available (as
we can have multiple local images)