Parse monument: Rework incremental parsing
Intially, incremental parsing was introduced to help with full parses, which
take several days and often fail. For some reason, the old code did not use the
incremental data on full parse.

We reworked the algorithm to simplify the code and use incremental data
regardless of the parse_type we choose.

Signed-off-by: Strainu <github@strainu.ro>