feat(localitati): Handle more methods to find images.
In addition to the existing code, we also recognize videos and image collages.
We use images from UAT's villages to illustrate the UAT. We also have a
disabled (for runtime sake) feature to search for articles in hu, de articles.
This proved useful in Transylvania, providing us with several tens of images.

Signed-off-by: Strainu <github@strainu.ro>